Break out custom bars
